hydroponics-system/requirements.txt
flask==3.0.0
RPi.GPIO==0.7.1
spidev==3.6
w1thermsensor==2.3.0
pyserial==3.5